Streamlin_requirements.txt
streamlit
langchain
langchain-core
langchain-chroma
langchain-google-genai
google-generativeai
langchain-community